+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 == Changelog ===
 
+= 5.11.1 =
+* Fix two vulnerabilities.
+
 = 5.11.0 =
 * Update embedded Matomo to version 5.11.1.
 * Bug fix: fix incorrect capability check that used current user in some cases instead of requested user.